% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcReplicateCorr.R
\name{calcReplicateCorr}
\alias{calcReplicateCorr}
\title{Calculate correlation between technical replicates}
\usage{
calcReplicateCorr(dge, group, corr.thresh = 0.8, return = "good")
}
\arguments{
\item{dge}{DGEList object containing technical replicates to be compared}
}
\value{
Returns a plot of the read counts per barcode (row) in a data frame
}
\description{
Calculate and return list of correlation between paired technical replicates in a dataset
}
\examples{
calcReplicateCor(dge, group = group)

}
