% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawBarcodeAnnoPlots.R
\name{drawBarcodeAnnoPlots}
\alias{drawBarcodeAnnoPlots}
\title{Draw Barcode Annotation Plots}
\usage{
drawBarcodeAnnoPlots(
  bam.parsed.df,
  breaks = 50,
  outdir = "./",
  prefix = "barcode_anno_plots"
)
}
\arguments{
\item{bam.parsed.df}{Dataframe from mapBarcodeReads() containing annotated barcode reads}

\item{breaks}{Optional. Bin size for histograms}

\item{outdir}{Optional. Output directory. Will generate a plots directory inside here to store plots}

\item{prefix}{Optional. Prefix to add to filename for each plot}
}
\value{
Returns a data-frame containing the read ID, the 10X cell ID, 10X umi, lintraceR DNA
barcode sequence mapped to the lintraceR DNA barcode ID.
}
\description{
Plot summary statistics resulting from 10X barcode annotation
}
\examples{
drawBarcodeAnnoPlots(bam.parsed.df, breaks = 50, outdir = "./", prefix = "barcode_anno_plots")
}
