% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nthPercentileBarcodes.R
\name{nthPercentile_barcodes}
\alias{nthPercentile_barcodes}
\title{nthPercentile_barcodes}
\usage{
nthPercentile_barcodes(counts, percentile = 0.95)
}
\arguments{
\item{percentile}{desired percentile value}

\item{counts.obj}{dataframe containing raw or normalised barcode counts}
}
\value{
Returns a cumulative sum & percentile plot per sample.
}
\description{
Calculate barcodes comprising the Nth percentile for each sample & generate cumulative sum plots
}
\examples{
nthPercentile_barcodes(counts, percentile = .95)
}
