% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeHeatmap.R
\name{plot_proportional_heatmap}
\alias{plot_proportional_heatmap}
\title{plot_proportional_heatmap}
\usage{
plot_proportional_heatmap(
  counts.obj,
  outfile,
  name = "Proportional Barcode Plot",
  scale = "row"
)
}
\arguments{
\item{counts.obj}{dataframe containing raw / normalised counts of barcodes}

\item{outfile}{desired name of output plot}

\item{name}{desired plot title}

\item{scale}{scaling to use for heatmap: can be none, row or column as per pheatmap}
}
\value{
Returns a plot of barcodes represented by proportion of total pool
}
\description{
Generate proportional heatmap from raw count / normalised count object.
}
\examples{
plot_proportional_heatmap(counts.obj, outfile, name = "Proportional Barcode Plot", scale = "row")
}
