% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeTimeseries.R
\name{plot_proportional_timeseries}
\alias{plot_proportional_timeseries}
\title{plot_proportional_timeseries}
\usage{
plot_proportional_timeseries(counts.obj, outfile, name = "", seed = 5)
}
\arguments{
\item{counts.obj}{dataframe containing raw counts of barcodes}

\item{outfile}{desired name of output plot}

\item{name}{desired plot title}

\item{seed}{RNG seed}
}
\value{
Returns a bubbleplot of barcodes represented by proportion of total pool
}
\description{
Generate proportional timeseries plot from raw / normalised barcode count object.
}
\examples{
plot_proportional_timeseries(counts.obj, name = "Proportional Bubble Plot", seed = 5)
}
