% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeBubble.R
\name{plotBarcodeBubble}
\alias{plotBarcodeBubble}
\title{proportionalBubbleplot}
\usage{
plotBarcodeBubble(
  counts.obj,
  labels = T,
  proportion.cutoff = 10,
  name = "Proportional Bubble Plot"
)
}
\arguments{
\item{counts.obj}{dataframe containing raw counts of barcodes}

\item{labels}{Boolean. print barcode labels?}

\item{proportion.cutoff}{barcodes represented at a percentage within any sample above this threshold will be labelled}

\item{name}{desired plot title}
}
\value{
Returns a bubbleplot of barcodes represented by proportion of total pool
}
\description{
Generate proportional bubbleplots from raw count object with barcodes labelled above a specified threshold
}
\examples{
proportionalBubbleplot(counts.obj, name = "Proportional Bubble Plot", label = T, proportion.cutoff = 10)
}
