% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeFrequency.R
\name{plotBarcodeFrequency}
\alias{plotBarcodeFrequency}
\title{Plot barcode frequency distribution}
\usage{
plotBarcodeFrequency(barcodes, samplename)
}
\arguments{
\item{barcodes}{DGEList or DESeqDataSet object containing raw or normalised barcode counts with replicate grouping info in object metadata}

\item{samplename}{a character vector of containing grouping information. Must be as long as the columns of object. Can pass a metadata column from object.}
}
\value{
Returns a plot of the barcode frequency distribution for a sample
}
\description{
Plot barcode frequency distribution from a two column data frame containing barcode name and count.
}
\examples{
plotBarcodeFrequency(barcodes, samplename = "test sample")

}
