% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeRegression.R
\name{plotBarcodeRegression}
\alias{plotBarcodeRegression}
\title{plotBarcodeRegression}
\usage{
plotBarcodeRegression(
  dge,
  group,
  title = paste("Regression plot:", group),
  rug = T,
  trendline = T,
  stats = T,
  trans = NULL
)
}
\arguments{
\item{dge}{DGEList object containing grouping variable to fit linear model}

\item{group}{sample group in dge$samples$group to plot correlation for. must be paired}

\item{title}{desired name of output plot}

\item{rug}{boolean. include geom_rug density information on the axes?}

\item{trendline}{boolean. include linear trendline using stat_smooth()?}

\item{stats}{boolean. include model stats in the plot subtitle?}

\item{trans}{From ggplot2. For continuous scales, the name of a transformation object or the object itself. Built-in transformations include "asn", "atanh", "boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2", "logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal", "reverse", "sqrt" and "time".}

\item{fit}{fit of linear model between two samples generated using lm}
}
\value{
Returns a scatterplot represented by proportion of total pool
}
\description{
Generate a linear regression scatterplot from a model fit using lm() between two sets of sample counts.
}
\examples{
plotBarcodeRegression(fit, title = "Proportional Bubble Plot", rug = T)
}
