% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeScatter.R
\name{plotBarcodeScatter}
\alias{plotBarcodeScatter}
\title{plotScatter}
\usage{
plotBarcodeScatter(
  obj = NULL,
  samps = NULL,
  title = "Scatter plot",
  rug = T,
  trendline = T,
  stats = T,
  trans = NULL
)
}
\arguments{
\item{obj}{DGEList object containing samples to be compared}

\item{samps}{character vector of samplenames of length 2 or more. more then 2 samples will create a pairwise scatter plot.}

\item{title}{desired name of output plot}

\item{rug}{boolean. include geom_rug density information on the axes?}

\item{trendline}{boolean. include linear trendline using stat_smooth()?}

\item{stats}{boolean. include model stats in the plot subtitle?}

\item{trans}{From ggplot2. For continuous scales, the name of a transformation object or the object itself. Built-in transformations include "asn", "atanh", "boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2", "logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal", "reverse", "sqrt" and "time".}
}
\value{
Returns a scatterplot between two or more sets of sample counts.
}
\description{
Generate a simple scatterplot between two or more sets of sample counts.
}
\examples{
plotScatter(obj, samps = c("Samp_1", "Samp_2"), title = "Scatter plot", rug = T, trendline = T, stats = T, trans = NULL)
}
