% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeDist.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Sample to sample correlation}
\usage{
plotCorrelation(
  counts,
  method = "pearson",
  upper = T,
  clustered = T,
  name = "Sample Correlation"
)
}
\arguments{
\item{counts.obj}{matrix/dataframe containing raw or normalised counts}
}
\value{
Returns a heatmap of sample distances using desired clustering
}
\description{
Plot sample correlation between barcode sets / samples
}
\examples{
test.mat <- matrix(rnorm(10*10,mean=0,sd=2), 10, 10)
plotCorrelation(test.mat, method = "pearson")

}
