% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeDist.R
\name{plotDistances}
\alias{plotDistances}
\title{Sample to sample distances}
\usage{
plotDistances(
  counts.obj,
  method = "euclidean",
  upper = T,
  name = "Sample Distances"
)
}
\arguments{
\item{counts.obj}{matrix/dataframe containing raw or normalised counts}
}
\value{
Returns a heatmap of sample distances using desired clustering
}
\description{
Plot sample distances between barcode sets / samples
}
\examples{
test.mat <- matrix(rnorm(10*10,mean=0,sd=2), 10, 10)
plotDistances(test.mat)

}
