% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBarcodeObj.R
\name{process_barcode_counts}
\alias{process_barcode_counts}
\title{createBarcodeObj}
\usage{
process_barcode_counts(counts.dir = "counts/", sampleanno)
}
\arguments{
\item{counts.dir}{directory containing counts files per sample}

\item{sampleanno}{annotation file detailing sample ID and groups}
}
\value{
A data frame containing barcode counts in rows per sample in columns.
}
\description{
generate a DGEList object containing barcode counts and sample metadata from raw counts files
}
\examples{
process_barcode_counts(counts.dir = "counts/", sampleanno)
}
