% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBarcodeLibrary.R
\name{processBarcodeLibrary}
\alias{processBarcodeLibrary}
\title{processBarcodeLibrary}
\usage{
processBarcodeLibrary(
  filepath = NULL,
  samplename = "Barcode",
  cutoff = 10,
  header = F
)
}
\arguments{
\item{filepath}{path to starcode output for barcode library}
}
\value{
returns data frame containing barcode and read count in reference
}
\description{
process barcode reference file from raw sequencing datasets
}
\examples{
process_barcode_reference(filepath = NULL)
}
